pyproject.toml
[project]
name = "scythe-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.40",
    "openai>=1.0.0",
]
